refactor: DashViewModelのインタフェースを削除 viewModelのインタフェースは必要ない
